test: consolidate on a single ethereum conformance suite
Retire the old ethereum/tests corpus tests (TestEthereumTests, TestSingleAdd11)
and the -legacy path, and fold the execution-specs state_tests suite in as the
sole TestEthereumTests. The shared harness (runStateTestSubtests,
runEthereumTestConfig, the priming/trie helpers) is unchanged; only the
old-only entry points and the -legacy flag are removed.

Rename the skip/slow lists to the plain names (execution_specs_tests.skip/.slow
-> eth_tests.skip/.slow) now that there is a single suite.

Refs #238
