+++ testdata/eth_tests.slow
@@ -1,57 +1,3 @@
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stPreCompiledContracts/precompsEIP2929Cancun.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Call50000_ecrec.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Call50000_identity2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Call50000_identity.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Call50000.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Callcode50000.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/QuadraticComplexitySolidity_CallDataCopy.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Return50000_2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stQuadraticComplexityTest/Return50000.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stRevertTest/LoopCallsThenRevert.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStackTests/underflowTest.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_Call50000_ecrec.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_Call50000_identity2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_Call50000_identity.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_Call50000.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_Call50000_rip160.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_LoopCallsThenRevert.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stStaticCall/static_Return50000_2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/CALLBlake2f_MaxRounds.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/sstore_combinations_initial0_2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/sstore_combinations_initial0.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/sstore_combinations_initial1_2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/sstore_combinations_initial1.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/sstore_combinations_initial2_1.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/sstore_combinations_initial2.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/stTimeConsuming/static_Call50000_sha256.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/VMTests/vmIOandFlowOperations/jumpToPush.json
-../testdata/ethereum-tests/LegacyTests/Constantinople/GeneralStateTests/VMTests/vmPerformance/loopMul.json
-
-../testdata/ethereum-tests/GeneralStateTests/stPreCompiledContracts/precompsEIP2929Cancun.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Call50000_ecrec.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Call50000_identity2.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Call50000_identity.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Call50000.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Callcode50000.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/QuadraticComplexitySolidity_CallDataCopy.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Return50000_2.json
-../testdata/ethereum-tests/GeneralStateTests/stQuadraticComplexityTest/Return50000.json
-../testdata/ethereum-tests/GeneralStateTests/stRevertTest/LoopCallsThenRevert.json
-../testdata/ethereum-tests/GeneralStateTests/stStackTests/underflowTest.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_Call50000_ecrec.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_Call50000_identity2.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_Call50000_identity.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_Call50000.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_Call50000_rip160.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_LoopCallsThenRevert.json
-../testdata/ethereum-tests/GeneralStateTests/stStaticCall/static_Return50000_2.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/CALLBlake2f_MaxRounds.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/sstore_combinations_initial0_2.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/sstore_combinations_initial0.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/sstore_combinations_initial1_2.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/sstore_combinations_initial1.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/sstore_combinations_initial2_1.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/sstore_combinations_initial2.json
-../testdata/ethereum-tests/GeneralStateTests/stTimeConsuming/static_Call50000_sha256.json
-../testdata/ethereum-tests/GeneralStateTests/VMTests/vmIOandFlowOperations/jumpToPush.json
-../testdata/ethereum-tests/GeneralStateTests/VMTests/vmPerformance/loopMul.json
+# Slow tests for the execution-specs conformance suite (run only with -very_slow).
+# One test path per line; blank lines and lines starting with '#' are ignored.
+# Populated in the skip/slow triage chunk (#237) — empty for now.
